migrate resource tests
