+++ ckanext/unhcr/tests/nose/test_actions.py
@@ -277,105 +277,6 @@ class TestDatastoreAuthRestrictedDownloads(base.FunctionalTestBase):
             resource_id=self.resource['id'])
 
 
-class TestResourceUpload(base.FunctionalTestBase):
-
-    def test_upload_present(self):
-
-        dataset = factories.Dataset()
-
-        resource = factories.Resource(
-            package_id=dataset['id'],
-            upload=mocks.FakeFileStorage(),
-            url = "http://fakeurl/test.txt",
-            url_type='upload',
-        )
-
-        assert_equals(
-            resource['url'],
-            '{}/dataset/{}/resource/{}/download/test.txt'.format(
-                toolkit.config.get('ckan.site_url').rstrip('/'),
-                dataset['id'],
-                resource['id']
-            )
-        )
-
-    def test_upload_present_after_update(self):
-
-        dataset = factories.Dataset()
-
-        resource = factories.Resource(
-            package_id=dataset['id'],
-            upload=mocks.FakeFileStorage(),
-            url = "http://fakeurl/test.txt",
-            url_type='upload',
-        )
-
-        resource['name'] = 'updated'
-        updated_resource = core_helpers.call_action(
-            'resource_update',
-            {'ignore_auth': True},
-            **resource
-        )
-
-        assert_equals(updated_resource['name'], 'updated')
-
-        assert_equals(
-            updated_resource['url'],
-            '{}/dataset/{}/resource/{}/download/test.txt'.format(
-                toolkit.config['ckan.site_url'].rstrip('/'),
-                dataset['id'],
-                resource['id']
-            )
-        )
-
-    def test_upload_external_url_data(self):
-
-        dataset = factories.Dataset()
-
-        with assert_raises(toolkit.ValidationError) as exc:
-
-            factories.Resource(
-                type='data',
-                package_id=dataset['id'],
-                url='https://example.com/some.data.csv'
-            )
-
-        assert exc.exception.error_dict.keys() == ['url']
-
-        assert_equals(
-            exc.exception.error_dict['url'],
-            ['All data resources require an uploaded file'])
-
-    def test_upload_external_url_attachment(self):
-        dataset = factories.Dataset()
-        resource = factories.Resource(
-            type='attachment',
-            package_id=dataset['id'],
-            url='https://example.com/some.data.csv',
-            file_type='other'
-        )
-
-        # tbh, the main thing we're testing here is that the line above
-        # runs without throwing a ValidationError
-        # but I supposed we should assert _something_
-        assert_equals(resource['url'], 'https://example.com/some.data.csv')
-
-    def test_upload_missing(self):
-
-        dataset = factories.Dataset()
-
-        with assert_raises(toolkit.ValidationError) as exc:
-
-            factories.Resource(package_id=dataset['id'])
-
-        assert exc.exception.error_dict.keys() == ['url']
-
-        assert_equals(
-            exc.exception.error_dict['url'],
-            ['All data resources require an uploaded file']
-        )
-
-
 class TestPendingRequestsList(base.FunctionalTestBase):
     def setup(self):
         super(TestPendingRequestsList, self).setup()
